% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{fuzz_all}
\alias{fuzz_all}
\alias{fuzz_char}
\alias{fuzz_dbl}
\alias{fuzz_fctr}
\alias{fuzz_int}
\alias{fuzz_log}
\title{Fuzz test inputs}
\usage{
fuzz_all()

fuzz_char()

fuzz_int()

fuzz_dbl()

fuzz_log()

fuzz_fctr()
}
\description{
Fuzz test inputs
}
\section{Functions}{
\itemize{
\item \code{fuzz_char}: Character vectors

\item \code{fuzz_int}: Integer vectors

\item \code{fuzz_dbl}: Double vectors

\item \code{fuzz_log}: Logical vectors

\item \code{fuzz_fctr}: Factor vectors
}}

