% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{fuzz_value}
\alias{fuzz_call}
\alias{fuzz_value}
\title{Access the object returned by the fuzz test}
\usage{
fuzz_value(fr, index)

fuzz_call(fr, index)
}
\arguments{
\item{fr}{fuzz_results object}

\item{index}{The test index (by position), or name, whose results to access.
Same as the row number in the data frame returned by
\code{\link{as.data.frame.fuzz_results}}.}
}
\description{
Access the object returned by the fuzz test
}
\section{Functions}{
\itemize{
\item \code{fuzz_call}: Access the call used for the fuzz test
}}

