% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{is.fuzz_results}
\alias{is.fuzz_results}
\title{Is the object of class fuzz_results?}
\usage{
is.fuzz_results(x)
}
\arguments{
\item{x}{}
}
\value{
Logical value
}
\description{
Is the object of class fuzz_results?
}

