% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{summary.fuzz_results}
\alias{summary.fuzz_results}
\title{Summarize fuzz test results as a data frame}
\usage{
\method{summary}{fuzz_results}(fuzz_results)
}
\arguments{
\item{fuzz_results}{object returned by \code{\link{fuzz_function}}.}
}
\value{
A data frame with the following columns: \describe{
\item{\code{fuzz_input}}{The name of the fuzz test performed.}
\item{\code{type}}{The type of response. If a condition was returned, one of
   \code{message}, \code{warning}, or \code{error}; if a value was returned,
   the \code{\link{class}} of that value.}
\item{\code{message}}{The condition message, if applicable.}
}
}
\description{
Summarize fuzz test results as a data frame
}

