% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{test_null}
\alias{test_null}
\title{describeIn test_all null \itemize{
 \item \code{null_value}: \code{NULL}
}}
\usage{
test_null()
}
\description{
describeIn test_all null \itemize{
 \item \code{null_value}: \code{NULL}
}
}

