% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{value_returned}
\alias{value_returned}
\title{Access the object returned by the fuzz test}
\usage{
value_returned(fuzz_results, index)
}
\arguments{
\item{fuzz_results}{fuzz_results object}

\item{index}{The test index (by position), or name, whose results to access. Same as the
row number in the data frame returned by
\code{\link{as.data.frame.fuzz_results}}.}
}
\description{
Access the object returned by the fuzz test
}

