% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{as.data.frame.fuzz_results}
\alias{as.data.frame.fuzz_results}
\title{Summarize fuzz test results as a data frame}
\usage{
\method{as.data.frame}{fuzz_results}(x, ..., delim = "; ",
  .id = "fuzz_input")
}
\arguments{
\item{x}{Object returned by \code{\link{fuzz_function}}.}

\item{...}{Additional arguments to be passed to or from methods.}

\item{delim}{The delimiter to use in fields like \code{messages} or
\code{warnings} where there may be multiple results.}

\item{.id}{Quoted column name for the name of the fuzz input used for each
test.}
}
\value{
A data frame with the following columns: \describe{
  \item{\code{fuzz_input}}{The name of the fuzz test performed.}
  \item{\code{output}}{Delimited outputs to the command line from the process, if applicable.}
  \item{\code{messages}}{Delimited messages, if applicable.}
  \item{\code{warnings}}{Delimited warnnings, if applicable.}
  \item{\code{errors}}{Error returned, if applicable.}
  \item{\code{value_classes}}{Delimited classes of the object returned by the
   function, if applicable}
  }
}
\description{
Summarize fuzz test results as a data frame
}

