% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluators.R
\name{fuzz_function}
\alias{fuzz_function}
\title{Fuzz-test a function}
\usage{
fuzz_function(fun, arg_name, ..., tests = test_all())
}
\arguments{
\item{fun}{A function, either bare or quoted.}

\item{arg_name}{Quoted name of the argument to fuzz test.}

\item{...}{Other non-dynamic values to pass to \code{fun}.}

\item{tests}{Which fuzz tests to run. Accepts a named list of inputs, defaulting to \code{\link{fuzz_all}}.}
}
\value{
A \code{fuzz_results} object.
}
\description{
Fuzz-test a function
}
\seealso{
\code{\link{as.data.frame.fuzz_results}} and \code{\link{value_returned}} to access fuzz test results.
}

