% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluators.R
\name{fuzz_function}
\alias{fuzz_function}
\title{Fuzz-test a function}
\usage{
fuzz_function(fun, arg_name, ..., tests = fuzz_all())
}
\arguments{
\item{fun}{A function}

\item{...}{Static values to pass to fun}

\item{tests}{Which fuzz tests to run. Accepts a named list of inputs, defaulting to \code{\link{fuzz_all}}.}

\item{arg_names}{Quoted name of the argument to fuzz test}
}
\value{
A data frame of fuzz test results.
}
\description{
Fuzz-test a function
}

