% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{fuzz_results}
\alias{fuzz_call}
\alias{fuzz_results}
\alias{fuzz_value}
\title{Access individual fuzz test results}
\usage{
fuzz_value(fr, index)

fuzz_call(fr, index)
}
\arguments{
\item{fr}{\code{fuzz_results} object}

\item{index}{The test index (by position) to access. Same as the
\code{results_index} in the data frame returned by
\code{\link{as.data.frame.fuzz_results}}.}
}
\description{
Access individual fuzz test results
}
\section{Functions}{
\itemize{
\item \code{fuzz_value}: Access the object returned by the fuzz test

\item \code{fuzz_call}: Access the call used for the fuzz test
}}

