% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzr-package.R
\docType{package}
\name{fuzzr}
\alias{fuzzr}
\alias{fuzzr-package}
\title{Fuzz-Test R Functions}
\description{
Test function arugments with a wide array of inputs, and produce reports
summarizing messages, warnings, errors, and returned values.
}
