% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{test_all}
\alias{test_all}
\alias{test_char}
\alias{test_dbl}
\alias{test_fctr}
\alias{test_int}
\alias{test_lgl}
\title{Fuzz test inputs}
\usage{
test_all()

test_char()

test_int()

test_dbl()

test_lgl()

test_fctr()
}
\description{
Fuzz test inputs
}
\section{Functions}{
\itemize{
\item \code{test_char}: Character vectors

\item \code{test_int}: Integer vectors

\item \code{test_dbl}: Double vectors

\item \code{test_lgl}: Logical vectors

\item \code{test_fctr}: Factor vectors
}}

