% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{test_all}
\alias{test_all}
\alias{test_char}
\alias{test_date}
\alias{test_dbl}
\alias{test_df}
\alias{test_fctr}
\alias{test_int}
\alias{test_lgl}
\alias{test_raw}
\title{Fuzz test inputs}
\usage{
test_all()

test_char()

test_int()

test_dbl()

test_lgl()

test_fctr()

test_date()

test_raw()

test_df()
}
\description{
Fuzz test inputs
}
\section{Functions}{
\itemize{
\item \code{test_char}: Character vectors

\item \code{test_int}: Integer vectors

\item \code{test_dbl}: Double vectors

\item \code{test_lgl}: Logical vectors

\item \code{test_fctr}: Factor vectors

\item \code{test_date}: Date vectors

\item \code{test_raw}: Raw vectors

\item \code{test_df}: Data frames
}}

